@@ -2,14 +2,14 @@
     "name": "query-builder",
     "version": "0.1.0",
     "description": "Query Builder typescript",
-    "main": "index.js",
     "scripts": {
         "clean": "rimraf lib",
-        "build": "rollup --config=./rollup.config.js",
-        "watch": "rollup --watch --config=./rollup.config.js",
+        "build": "rollup --config=./rollup.config.ts --configPlugin typescript",
+        "watch": "rollup --watch --config=./rollup.config.ts --configPlugin typescript",
         "lint": "eslint -c .eslintrc.cjs src/**/*.{ts,tsx,json}",
         "lint:fix": "eslint -c .eslintrc.cjs --fix src/**/*.{ts,tsx,json}",
-        "format": "prettier --write src/**/*.{ts,tsx,json} --config ./.prettierrc"
+        "format": "prettier --write src/**/*.{ts,tsx,json} --config ./.prettierrc",
+        "test": "jest"
     },
     "repository": {
         "type": "git",
@@ -28,22 +28,30 @@
     },
     "homepage": "https://github.com/joalvm/query-builder#readme",
     "devDependencies": {
-        "@rollup/plugin-commonjs": "^24.0.1",
-        "@rollup/plugin-terser": "^0.4.0",
-        "@rollup/plugin-typescript": "^11.0.0",
-        "@typescript-eslint/eslint-plugin": "^5.53.0",
-        "@typescript-eslint/parser": "^5.53.0",
-        "eslint": "^8.34.0",
-        "eslint-config-prettier": "^8.6.0",
-        "eslint-import-resolver-typescript": "^3.5.3",
-        "eslint-plugin-import": "^2.27.5",
-        "eslint-plugin-prettier": "^4.2.1",
+        "@babel/preset-env": "^7.22.20",
+        "@babel/preset-typescript": "^7.23.0",
+        "@rollup/plugin-commonjs": "^25.0.4",
+        "@rollup/plugin-terser": "^0.4.3",
+        "@rollup/plugin-typescript": "^11.1.4",
+        "@types/jest": "^29.5.5",
+        "@typescript-eslint/eslint-plugin": "^6.7.3",
+        "@typescript-eslint/parser": "^6.7.3",
+        "babel-jest": "^29.7.0",
+        "eslint": "^8.50.0",
+        "eslint-config-prettier": "^9.0.0",
+        "eslint-import-resolver-typescript": "^3.6.1",
+        "eslint-plugin-import": "^2.28.1",
+        "eslint-plugin-prettier": "^5.0.0",
         "eslint-plugin-simple-import-sort": "^10.0.0",
-        "prettier": "^2.8.4",
-        "rimraf": "^4.1.2",
-        "rollup": "^3.17.2",
-        "tslib": "^2.5.0",
-        "typescript": "^4.9.5"
+        "eslint-plugin-unused-imports": "^3.0.0",
+        "jest": "^29.7.0",
+        "prettier": "^3.0.3",
+        "rimraf": "^5.0.5",
+        "rollup": "^3.29.4",
+        "ts-jest": "^29.1.1",
+        "ts-node": "^10.9.1",
+        "tslib": "^2.6.2",
+        "typescript": "^5.2.2"
     },
     "exports": {
         ".": {
