@@ -1,30 +1,30 @@
-const { defineConfig } = require("rollup");
-const tsc = require("@rollup/plugin-typescript");
-const commonjs = require("@rollup/plugin-commonjs");
-const terser = require("@rollup/plugin-terser");
+import { defineConfig } from 'rollup';
+import tsc from '@rollup/plugin-typescript';
+import commonjs from '@rollup/plugin-commonjs';
+import terser from '@rollup/plugin-terser';
 
-const input = "./src/query.ts";
+const input = './src/builder.ts';
 
-exports.default = [
+export default [
     defineConfig({
-        context: "./",
+        context: './',
         input,
         output: [
             {
                 file: `./lib/esm/query.js`,
-                format: "esm",
+                format: 'esm',
                 compact: true,
                 sourcemap: true,
             },
             {
                 file: `./lib/esm/query.min.js`,
-                format: "esm",
+                format: 'esm',
                 compact: true,
                 sourcemap: true,
                 plugins: [terser()],
             },
         ],
-        plugins: [tsc({ tsconfig: "./configs/tsconfig.esm.json" })],
+        plugins: [tsc({ tsconfig: './configs/tsconfig.esm.json' })],
     }),
     defineConfig({
         input,
@@ -32,39 +32,36 @@ exports.default = [
             {
                 file: `./lib/cjs/query.js`,
                 compact: true,
-                format: "cjs",
+                format: 'cjs',
             },
             {
                 file: `./lib/cjs/query.min.js`,
-                format: "cjs",
+                format: 'cjs',
                 compact: true,
                 plugins: [terser()],
             },
         ],
-        plugins: [
-            tsc({ tsconfig: "./configs/tsconfig.cjs.json" }),
-            commonjs({ extensions: [".js", ".ts"] }),
-        ],
+        plugins: [tsc({ tsconfig: './configs/tsconfig.cjs.json' }), commonjs({ extensions: ['.js', '.ts'] })],
     }),
     defineConfig({
         input,
         output: [
             {
                 file: `./lib/umd/query.js`,
-                format: "umd",
-                name: "Query",
+                format: 'umd',
+                name: 'Query',
                 compact: true,
                 sourcemap: true,
             },
             {
                 file: `./lib/umd/query.min.js`,
-                format: "umd",
-                name: "Query",
+                format: 'umd',
+                name: 'Query',
                 compact: true,
                 sourcemap: true,
                 plugins: [terser()],
             },
         ],
-        plugins: [tsc({ tsconfig: "./configs/tsconfig.umd.json" })],
+        plugins: [tsc({ tsconfig: './configs/tsconfig.umd.json' })],
     }),
 ];
